% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{bd.gwas.test}
\alias{bd.gwas.test}
\title{Fast K-sample Ball Divergence Test for GWAS Data}
\usage{
bd.gwas.test(x, snp, screening.method = c("permute", "spectrum"),
  num.permutations, distance = FALSE, alpha, verbose = TRUE,
  seed = 1, num.threads = 0, ...)
}
\arguments{
\item{x}{a numeric vector, matrix, data.frame, dist object}

\item{snp}{SNP data}

\item{screening.method}{if \code{screening.method = "spectrum"}, the spectrum method is applied to 
screening the candidate SNPs, or otherwise, the permutation method is applied.}

\item{num.permutations}{the number of permutation replications, when \code{num.permutations} equals to 0, the function returns
the sample version of ball divergence. Default: \code{num.permutations = 99}.}

\item{distance}{if \code{distance = TRUE}, the elements of \code{x} will be considered as a distance matrix. Default: \code{distance = FALSE}.}

\item{alpha}{the significance level.}

\item{verbose}{Show computation status and estimated runtime. Default: \code{verbose = FALSE}.}

\item{seed}{the random seed. Default \code{seet = 1}.}

\item{num.threads}{Number of threads. If \code{num.threads = 0}, then all of available cores will be used. Default \code{num.threads = 0}.}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
bd.gwas.test returns a list containing the following components:
\item{\code{statistic}}{ball divergence statistics vector.}            
\item{\code{permuted.statistic}}{a data.frame containing permuted ball divergence statistic.}            
\item{\code{p.value}}{the p-values for the ball divergence test.}
}
\description{
Fast K-sample Ball Divergence Test for GWAS Data
}
\examples{
\dontrun{

set.seed(1)
num <- 200
snp_num <- 10000
p <- 5
x <- matrix(rnorm(num * p), nrow = num)
snp <- sapply(1:snp_num, function(i) {
  sample(0:2, size = num, replace = TRUE)
})
snp1 <- sapply(1:snp_num, function(i) {
  sample(1:2, size = num, replace = TRUE)
})
snp <- cbind(snp, snp1)
res <- Ball::bd.gwas.test(x = x, snp = snp, num.permutations = 19999)
mean(res[["p.value"]] < 0.05)
mean(res[["p.value"]] < 0.005)
mean(res[["p.value"]] < 0.0005)

num <- 100
snp_num <- 500
x <- rbind(matrix(rnorm(num * p), nrow = num), 
           matrix(rnorm(num * p, mean = 3), nrow = num), 
           matrix(rnorm(num * p, mean = -3), nrow = num))
snp <- sapply(1:snp_num, function(i) {
  sample(0:2, size = 3 * num, replace = TRUE)
})
snp[, 300] <- rep(0:2, each = num)
res <- Ball::bd.gwas.test(x = x, snp = snp)

num <- 200
snp_num <- 10000
x <- matrix(rnorm(num * p), nrow = num)
snp <- sapply(1:snp_num, function(i) {
  sample(0:2, size = num, replace = TRUE, prob = c(1/2, 1/4, 1/4))
})
res <- Ball::bd.gwas.test(x = x, snp = snp, alpha = 5*10^-4, num.permutations = 19999)
mean(res[["p.value"]] < 0.05)
mean(res[["p.value"]] < 0.005)
mean(res[["p.value"]] < 0.0005)
}
}
\seealso{
\code{\link{bd}}, \code{\link{bd.test}}
}
