% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_bcov.R
\name{KBCov}
\alias{KBCov}
\title{A New KBCOV Statistic}
\usage{
KBCov(X, n, method = "limit")
}
\arguments{
\item{method}{}
}
\description{
A New KBCOV Statistic
}
\examples{
library(Ball)
set.seed(1)
K=2
X <- list()
Y <- list()
n=1000
for(i in 1:K){
  Y[[i]] = rnorm(n)
  X[[i]] = as.vector(dist(Y[[i]]))
}
# when sample size is large the two statistics 
# have a similar statistic and test results.
KBCov(X, n)[["p.value"]]
bcov.test(Y[[1]], Y[[2]])

K <- 5
for(i in 1:K){
  Y[[i]] <- rnorm(n)
  X[[i]] <- as.vector(dist(Y[[i]]))
}
summary(KBCov(X, n)[["eigenvalues"]])
}
