% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcov_tmp.R
\name{KBCov_permuted}
\alias{KBCov_permuted}
\title{A New KBCOV Statistic}
\usage{
KBCov_permuted(X, n, num.permutations = 299, seed = 1)
}
\arguments{
\item{num.permutations}{199}

\item{seed}{1}
}
\value{

}
\description{
A New KBCOV Statistic
}
\examples{
set.seed(1)
K=2
X <- list()
Y <- list()
n=1000
for(i in 1:K){
  Y[[i]] = rnorm(n)
  X[[i]] = as.vector(dist(Y[[i]]))
}
# when sample size is large the two statistics 
# have a similar statistic and test results.
KBCov_permuted(X, n)
KBCov(X, n)
bcov.test(Y[[1]], Y[[2]])

}
