% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbcov_new.R
\name{kbcov_R}
\alias{kbcov_R}
\title{Title}
\usage{
kbcov_R(X)
}
\value{

}
\description{
Title
}
\examples{
K=2
n=200
x <- list(rnorm(n), rnorm(n))
bcov(x[[1]], x[[2]])
dst_x <- list()
for(i in 1:K){ dst_x[[i]] = as.vector(dist(x[[i]])) }
kbcov_R(dst_x)
}
