% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcov.R
\name{KBCovLimit}
\alias{KBCovLimit}
\title{A spectrum-based metric assocaition test (aka, Ball covariance test)}
\usage{
KBCovLimit(x)
}
\arguments{
\item{x}{a list containing at least two numeric distance matrices.}
}
\value{
\code{KBCovLimit} returns a list containing the following components:
\item{\code{statistic}}{Ball Covariance statistic.}            
\item{\code{p.value}}{the p-value for the test.}
}
\description{
A spectrum-based metric assocaition test (aka, Ball covariance test)
}
\examples{
library(Ball)
set.seed(1)
K <- 2
x <- list()
Y <- list()
n <- 200
for(i in 1:K){
  Y[[i]] <- rnorm(n)
  x[[i]] <- as.vector(dist(Y[[i]]))
}
KBCovLimit(x, n)[c("statistics", "p.value")]
summary(KBCovLimit(x, n)[["eigenvalues"]])
bcov.test(Y[[1]],Y[[2]], method = "limit")

set.seed(1)
x <- list()
Y <- list()
n <- 500
K <- 3
for(i in 1:K){
  Y[[i]] <- rnorm(n)
  x[[i]] <- as.vector(dist(Y[[i]]))
}
kbcov_res <- KBCovLimit(x, n)
kbcov_res[c("statistics", "p.value")]
summary(kbcov_res[["eigenvalues"]])
}
