% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{macaques}
\alias{macaques}
\title{Male and Female macaque data}
\format{
\itemize{
\code{macaques$x}: An array of dimension \eqn{7 \times 3 \times 18}

\code{macaques$group}: A factor indicating the sex ('m' for male and 'f' for female)
}
}
\description{
Male and female macaque skull data. 7 landmarks in 3 dimensions, 
18 individuals (9 males, 9 females)
}
\details{
In an investigation into sex differences in the crania of a species of 
Macaca fascicularis (a type of monkey), random samples of 9 male and 9 female 
skulls were obtained by Paul O’Higgins (Hull-York Medical School) (Dryden and Mardia 1993). 
A subset of seven anatomical landmarks was located on each cranium and the three-dimensional (3D) 
coordinates of each point were recorded.
}
\note{
Dryden, I.L. and Mardia, K.V. (1998). Statistical Shape Analysis, Wiley, Chichester.
}
\references{
Dryden, I. L. and Mardia, K. V. (1993). Multivariate shape analysis. Sankhya Series A, 55, 460-480.
}
