% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbd.R
\name{sub.dist}
\alias{sub.dist}
\title{Title}
\usage{
sub.dist(
  x,
  y,
  dist.fun = NULL,
  x.sub = NULL,
  y.sub = NULL,
  sub.scale = c("log2", "log", "n/log", "sqrt"),
  seed = 1
)
}
\arguments{
\item{x}{}

\item{y}{}

\item{dist.fun}{}

\item{x.sub}{}

\item{y.sub}{}

\item{sub.scale}{}
}
\value{

}
\description{
Title
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
sub.dist(x, y)
}
