% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{bd.gwas.test}
\alias{bd.gwas.test}
\title{Fast Two Ball Divergence test for gwas data}
\usage{
bd.gwas.test(x, snp, num.permutations, distance = FALSE, seed = 1,
  num.threads = 0, ...)
}
\arguments{
\item{x}{a numeric vector, matrix, data.frame, dist object}

\item{snp}{SNP data}

\item{num.permutations}{the number of permutation replications, when \code{num.permutations} equals to 0, the function returns
the sample version of ball divergence. Default: \code{num.permutations = 99}}

\item{distance}{if \code{distance = TRUE}, \code{x} will be considered as a distance matrix. Default: \code{distance = FALSE}}

\item{seed}{the random seed.}

\item{num.threads}{Number of threads. Default \code{num.threads = 1}.}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
bd.gwas.test returns a list containing the following components:
\item{\code{statistic}}{ball divergence statistic.}            
\item{\code{permuted_statistic}}{a data.frame containing permuted ball divergence statistic.}            
\item{\code{p.value}}{the p-value for the test.}
}
\description{
Fast Two Ball Divergence test for gwas data
}
\examples{
set.seed(1)
num <- 200
snp_num <- 10000
p <- 5
x <- matrix(rnorm(num * p), nrow = num)
snp <- sapply(1:snp_num, function(i) {
  sample(0:2, size = num, replace = TRUE)
})
res <- bd.gwas.test(x = x, snp = snp, num.permutations = 29999, num.threads = 1)
mean(res[["p.value"]] < 0.05)
mean(res[["p.value"]] < 0.005)
mean(res[["p.value"]] < 0.0005)
}
