% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bd.gaws.R
\name{bd.gwas.test}
\alias{bd.gwas.test}
\title{Fast Two Ball Divergence test for gwas data}
\usage{
bd.gwas.test(x, snp, dst = FALSE, seed = 4, num.threads = 2, ...)
}
\arguments{
\item{x}{a numeric vector, matrix, data.frame, dist object}

\item{snp}{SNP data}

\item{dst}{if \code{dst = TRUE}, x will be considered as a distance matrix. Default: \code{dst = FALSE}}

\item{seed}{the random seed.}

\item{num.threads}{Number of threads. Default \code{num.threads = 2}.}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
bd.gwas.test returns a list containing the following components:
\item{\code{statistic}}{ball divergence statistic.}            
\item{\code{permuted_statistic}}{a data.frame containing permuted ball divergence statistic.}            
\item{\code{p.value}}{the p-value for the test.}
}
\description{
Fast Two Ball Divergence test for gwas data
}
\examples{
set.seed(1)
num <- 60
snp_num <- 100
p <- 10
x1 <- matrix(rnorm(num * p / 2), nrow = num/2)
x2 <- matrix(rnorm(num * p / 2, 3), nrow = num/2)
x <- rbind(x1, x2)
snp <- sapply(1:snp_num, function(i) { 
  set.seed(i)
  sample(c(0, 1), size = num, replace = TRUE)
})
snp[, 1] <- rep(c(0, 1), each = num/2)
snp[, 2] <- c(rep(0, num/2 - 1), rep(1, num/2 + 1))
snp[, 3] <- c(rep(0, num/2 - 2), rep(1, num/2 + 2))
snp[, 4] <- c(rep(0, num/2 - 4), rep(1, num/2 + 4))
res <- bd.gwas.test(x = x, snp = snp, num.threads = 2)
head(res[["p.value"]], n = 8)
}
